Move get_last_loc to the NPU allocator
Relocate the DSV4 table lookup helper into the shared NPU allocator module without changing its signature, body, comments, docstring, or call behavior. Import the helper from the concrete NPU module while leaving the main generic lookup untouched.

Extraction: awk '/^def get_last_loc\(/ {copying=1} copying && /^def alloc_paged_token_slots_extend_npu/ {exit} copying {print}' python/sglang/srt/hardware_backend/npu/dsv4/dsv4_allocator.py > /private/tmp/op30-wave2-0828de37-get_last_loc.before.py

Patch: apply_patch < /private/tmp/op30-wave2-0828de37-mechanical.patch

Verification: cmp /private/tmp/op30-wave2-0828de37-get_last_loc.before.py /private/tmp/op30-wave2-0828de37-get_last_loc.after.py
